feat: add effector localStorage
Remove utility functions for management local storage.
Rename file for effector.
