+++ package.json
@@ -11,6 +11,7 @@
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.1",
     "effector": "^22.8.6",
+    "effector-localstorage": "^0.5.3",
     "effector-react": "^22.5.3",
     "gh-pages": "^5.0.0",
     "react": "^18.2.0",
