+++ src/store/index.ts
@@ -1,3 +1,3 @@
-import { createTodoList } from "./effector";
+import { createTodoList } from "./model";
 
 export const model = createTodoList();
